LPS-99142 inline it unless we will reuse it. And if so, put it portal-test, not in portal-impl
